Fix custom model with numbers greater than 65535 dont work

+++ xLights/models/Node.h
@@ -43,8 +43,8 @@ public:
     };
 
     uint32_t ActChan = 0;   // 0 is the first channel
-    uint16_t sparkle = 0;
-    uint16_t StringNum = 0; // node is part of this string (0 is the first string)
+    uint32_t sparkle = 0;
+    uint32_t StringNum = 0; // node is part of this string (0 is the first string)
     std::vector<CoordStruct> Coords;
     std::string *name = nullptr;
     const Model *model = nullptr;
